Handle non-installed gems
When there are production gems that are not installed in development, graph
will not be able to find the gem.

Before
It threw an exception without any actionable information.

After
It still generates a graph, but also provides information so the user can exclude the group with -W if desired